backend da tela de busca
