+++ Mobile/rental_app/lib/screens/edit_immobile_screen.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import '../models/owner.dart';
+import '../models/immobile.dart';
 import '../services/api_service.dart';
 
 class EditImmobileScreen extends StatefulWidget {
