+++ Mobile/rental_app/lib/screens/owner_immobiles_screen.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import '../models/owner.dart'; // Immobile está dentro de owner.dart
+import '../models/immobile.dart';
 import '../services/api_service.dart';
 import 'edit_immobile_screen.dart';
